Move DocumentFactory to composition root to fix dependency inversion violation
DocumentFactory was in the application layer (com.embervault.application) but
depended on concrete adapter classes (InMemoryNoteRepository, etc.), violating
the Dependency Inversion Principle. Moved it to com.embervault where App.java
lives, since both serve as composition roots that wire abstractions to
implementations. Added an ArchUnit rule to prevent application-layer classes
from depending on adapter packages.

Closes #163
