+++ src/test/java/com/embervault/application/DocumentFactoryTest.java
@@ -8,6 +8,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.util.List;
 import java.util.Optional;
 
+import com.embervault.DocumentFactory;
 import com.embervault.domain.Note;
 import com.embervault.domain.Project;
 import org.junit.jupiter.api.DisplayName;
